+++ CHANGELOG.md
@@ -46,6 +46,7 @@
 * fix: [iOS] query `isHolding` only with no event emitted
 * feat: [iOS] CallKit controller used for mute / hold call actions
 * fix: [iOS] unable to disable call logging on iOS due to incorrect parameter from method channel args. 
+* fix: [iOS] use registered notification delegate instead of setting `UNUserNotificationCenter.current().delegate = self` to avoid overwriting other notification delegates.
 * Feat: update example.
 * Docs: update CHANGELOG